+++ src/opto_eq/__init__.py
@@ -1,4 +1,5 @@
 from .phase_modulator import PhaseModulator
+from .mzm import MZM
 from .optics import (coupler_split, coupler_combine, polarizer, polarization_rotator,
                      polarization_controller, pbs, beam_combiner, quarterwave, halfwave)
 from .fiber import cable